requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
adafruit-blinka==8.56.0
    # via
    #   adafruit-circuitpython-busdevice
    #   adafruit-circuitpython-connectionmanager
    #   adafruit-circuitpython-dht
    #   adafruit-circuitpython-requests
    #   adafruit-circuitpython-typing
adafruit-circuitpython-busdevice==5.2.11
    # via adafruit-circuitpython-typing
adafruit-circuitpython-connectionmanager==3.1.3
    # via adafruit-circuitpython-requests
adafruit-circuitpython-dht==4.0.7
    # via -r requirements.in
adafruit-circuitpython-requests==4.1.10
    # via adafruit-circuitpython-typing
adafruit-circuitpython-typing==1.11.2
    # via
    #   adafruit-blinka
    #   adafruit-circuitpython-busdevice
adafruit-platformdetect==3.77.0
    # via adafruit-blinka
adafruit-pureio==1.1.11
    # via adafruit-blinka
binho-host-adapter==0.1.6
    # via adafruit-blinka
build==1.2.2.post1
    # via pip-tools
click==8.1.8
    # via pip-tools
packaging==24.2
    # via build
paho-mqtt==2.1.0
    # via -r requirements.in
pip-tools==7.4.1
    # via -r requirements.in
pyftdi==0.56.0
    # via adafruit-blinka
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pyserial==3.5
    # via
    #   binho-host-adapter
    #   pyftdi
pyusb==1.3.1
    # via pyftdi
rpi-gpio==0.7.1
    # via -r requirements.in
schedule==1.2.2
    # via -r requirements.in
sysv-ipc==1.1.0
    # via adafruit-blinka
toml==0.10.2
    # via adafruit-blinka
tomli==2.2.1
    # via
    #   build
    #   pip-tools
typing-extensions==4.13.1
    # via adafruit-circuitpython-typing
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
